Move CryptException to util package
