+++ NachoSpigot-Server/src/main/java/me/elier/nachospigot/config/NachoConfig.java
@@ -15,7 +15,6 @@ import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
-import java.nio.file.Files;
 import java.util.List;
 
 public class NachoConfig {
